requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Pillow==7.2.0
websockets==8.1